secret_garden_service/pyproject.toml
[tool.poetry]
name = "secret-garden"
version = "0.1.0"
description = ""
authors = ["Antonio Moura"]
readme = "README.md"
packages = [{include = "secret_garden_service", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.115.11,<0.116.0"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.28"
pydantic-settings = "^2.8.1"
tabulate = "^0.9.0"
taskipy = ">=1.14.1,<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.3.7"
httpx = "^0.27.0"
isort = "^5.13.2"
factory-boy = "^3.3.0"
freezegun = "^1.4.0"
newrelic= "9.3.0"

[tool.taskipy.tasks]
lint = 'ruff check . && blue --check . --diff && isort --check . --diff'
format = 'blue .  && isort .'
run = 'uvicorn secret_garden_service.app:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=secret_garden_service -vv'
post_test = 'coverage html'

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['E402', 'F811']

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
